+++ panda/src/audio/audio_trait.h
@@ -46,7 +46,7 @@ public:
 
     virtual void play_sound(SoundClass*, PlayingClass*) = 0;
     virtual void stop_sound(SoundClass*, PlayingClass*) = 0;
-    virtual void set_volume(PlayingClass*, int) = 0;
+    virtual void set_volume(PlayingClass*, float) = 0;
   };
 };
